requirements.txt
# Core Dependencies
streamlit>=1.28.0
pydantic>=2.0.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database & Storage
chromadb>=0.4.0
neo4j>=5.0.0
pandas>=2.0.0

# NLP & ML
spacy>=3.7.0
ollama>=0.1.0
tiktoken>=0.5.0

# Document Processing
markitdown[all]>=0.0.1
python-docx>=0.8.11
pdfminer.six>=20221105
markdown>=3.5.0

# Development & Testing
pytest>=7.4.0
tqdm>=4.66.0
